아이디 권한 부여하기

+++ Server/controller/tweet.js
@@ -10,7 +10,7 @@ export async function getTweets(req, res) {
 
 export async function getTweet(req, res, next) {
     const id = req.params.id;
-    const tweet = await tweetRepository.geyByID(id);
+    const tweet = await tweetRepository.getById(id);
     if (tweet) {
         res.status(200).json(tweet);
     } else {
@@ -19,9 +19,9 @@ export async function getTweet(req, res, next) {
 }
 
 export async function createTweet(req, res, next) {
-    const { text, name, username } = req.body;
+    const { text } = req.body;
     // req.body에서 가져오기
-    const tweet = await tweetRepository.create(text, name, username)
+    const tweet = await tweetRepository.create(text, req.userId);
     res.status(201).json(tweet)
 
 }
@@ -29,18 +29,31 @@ export async function createTweet(req, res, next) {
 export async function updateTweet(req, res, next) {
     const id = req.params.id;
     const text = req.body.text;
-    const tweet = await tweetRepository.update(id, text);
-    if (tweet) {
-        res.status(200).json(tweet);
-    } else {
+    const tweet = await tweetRepository.getById(id);
+    // update 와 delete 에 특정토큰만 접근가능하게 만들기
+    if (!tweet) {
         res.status(404).json({ message: `tweet id(${id})not found` })
     }
+    if (tweet.userId !== req.userId) {
+        return res.sendStatus(403);
+    }
+
+    const updated = await tweetRepository.update(id,text);
+    res.status(200).json(updated);
 }
 
-export async function deleteTweet(req, res, next) {
-    const id = req.params.id
-    const tweets = await tweetRepository.remove(id)
-    res.sendStatus(204);
-}
+// update 와 delete 에 특정토큰만 접근가능하게 만들기
 
 // id로 설정한것 빼고 나머지를 선택한다
+export async function deleteTweet(req, res, next) {
+    const id = req.params.id
+    const tweet = await tweetRepository.getById(id);
+    if(!tweet) {
+        res.status(404).json({ message: `tweet id(${id})not found` })
+    }
+    if(usreId !== req.userID){
+        return res.sendStatus(403);
+    }
+    await tweetRepository.remove(id)
+    res.sendStatus(204);
+}
